모듈 시스템 초기 구현 및 로그 수정

+++ Engine/Source/Runtime/Core/String/StringTable.cpp
@@ -149,7 +149,7 @@ void GStringTable::removeOldStringInfos(int32 removeCountPerFrame)
 	if (removeCountPerFrame < 0)
 	{
 		bIsFlush = true;
-		removeCountPerFrame = _stringInfoMap.size();
+		removeCountPerFrame = (int32)_stringInfoMap.size();
 	}
 
 	int32 tempCnt = 0;
